Extra Auth and Style
Added in some basic style to the pages to have a nav bar that displays
the user information when they are logged in. Moved the static files
around so they are served correctly when no one is logged in. Basic
layout is set, will need to look into user roles and how to allow access
by tier then will start adding a user profile which will allow the user
to update all of their details