Reorganising and updating profiles
Moved and updated some files with better calls and methods to allow
consistent updates to user profiles. Added extra menu items which will
need to be gated by roles so need to start using the role assignment.
Index page needs to be created still profile is done for now as it
displays all the data and allows for updates